requirements.txt
Django==4.2
django-debug-toolbar==4.0.0
django-rest-swagger==2.2.0
djangorestframework==3.15.2
django-redis
Faker==26.0.0
python-dotenv==0.16.0
pytest-django==4.8.0
psycopg2-binary==2.9.1
